+++ src/models/kerasTF/models.py
@@ -43,7 +43,7 @@ LABELS_COLUMN_NAME = 'labels'
 os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
 warnings.filterwarnings('ignore')
 
-class KerasTFModels(ModelsUtils):
+class KerasTFModels(ModelsUtils, ABC):
     """
     Class used to build, train and predict models using Ray with Keras Tensorflow backend
 
@@ -94,247 +94,31 @@ class KerasTFModels(ModelsUtils):
             kmers_list,
             csv
         )
-        # Parameters
-        # Initialize hidden
-        self._nb_CPU_data = int(os.cpu_count() * 0.2)
-        self._nb_CPU_training = int(os.cpu_count() - self._nb_CPU_data)
-        self._nb_GPU = len(tf.config.list_physical_devices('GPU'))
-        # Initialize empty
-        self._nb_classes = None
-        self._nb_CPU_per_worker = 0
-        self._nb_GPU_per_worker = 0
-        # Computing variables
-        if self._nb_GPU > 0:
-            self._use_gpu = True
-            self._n_workers = self._nb_GPU
-            self._nb_CPU_per_worker = int(self._nb_CPU_training / self._n_workers)
-            self._nb_GPU_per_worker = 1
-        else:
-            self._use_gpu = False
-            self._n_workers = int(self._nb_CPU_training * 0.2)
-            self._nb_CPU_per_worker = int(int(self._nb_CPU_training * 0.8) / self._n_workers)
-
-        if self.classifier == 'attention':
-            print('Training bacterial / host classifier based on Attention Weighted Neural Network')
-        elif self.classifier == 'lstm':
-            print('Training bacterial / host classifier based on Shallow LSTM Neural Network')
-        elif self.classifier == 'deeplstm':
-            print('Training bacterial / host classifier based on Deep LSTM Neural Network')
-        elif self.classifier == 'lstm_attention':
-            print('Training multiclass classifier based on Deep Neural Network hybrid between LSTM and Attention')
-        elif self.classifier == 'cnn':
-            print('Training multiclass classifier based on CNN Neural Network')
-        elif self.classifier == 'widecnn':
-            print('Training multiclass classifier based on Wide CNN Network')
-
-    def preprocess(self, ds, scaling = False, scaler_file = None):
-        print('preprocess')
-        self._encoder = ModelLabelEncoder(self.taxa)
-        if scaling:
-            self._scaler = TensorTfIdfTransformer(self.kmers, scaler_file)
-            self._scaler.fit(ds)
-        self._encoder.fit(ds)
-        labels = list(self._encoder.stats_[f'unique_values({self.taxa})'].keys())
-        self._nb_classes = len(self._encoder.stats_[f'unique_values({self.taxa})'])
-        if self._nb_classes > 2 :
-            self._encoder = Chain(
-                self._encoder,
-                OneHotTensorEncoder(self.taxa)
-            )
-            self._encoder.fit(ds)
         
-        self._encoded = np.arange(len(labels))
-        labels = np.append(labels, 'Unknown')
-        self._encoded = np.append(self._encoded, -1)
-        for (label, encoded) in zip(labels, self._encoded):
-            self._labels_map[label] = encoded
-        self._compute_weights()
-
-    def _label_decode(self, predict):
-        print('_label_decode')
-        decoded = pd.Series(np.empty(len(predict), dtype=object))
-        for label, encoded in self._labels_map.items():
-            decoded[predict == encoded] = label
-
-        return np.array(decoded)
-
+    @abstractmethod
+    def preprocess(self):
+        """
+        """
+        
+    @abstractmethod
     def fit(self, datasets):
-        print('fit')
-        # Preprocessing loop
-        for name, ds in datasets.items():
-            ds = ds.drop_columns(['id'])
-            ds = self._encoder.transform(ds)
-            if self._scaler is not None:
-                ds = self._scaler.transform(ds)
-            # Trigger the preprocessing computations before ingest in trainer
-            # Otherwise, it would be executed at each epoch
-            ds = ds.materialize()
-            datasets[name] = ds
+        """
+        """
 
-        # Training parameters
-        self._train_params = {
-            'batch_size': self.batch_size,
-            'epochs': self._training_epochs,
-            'size': self._nb_kmers,
-            'nb_cls': self._nb_classes,
-            'model': self.classifier,
-            'weights': self._weights
-        }
-
-        # Define trainer / tuner
-        self._trainer = TensorflowTrainer(
-            train_loop_per_worker=train_func,
-            train_loop_config=self._train_params,
-            scaling_config=ScalingConfig(
-                trainer_resources={'CPU': self._nb_CPU_data},
-                num_workers=self._n_workers,
-                use_gpu=self._use_gpu,
-                resources_per_worker={
-                    'CPU': self._nb_CPU_per_worker,
-                    'GPU': self._nb_GPU_per_worker
-                }
-            ),
-            run_config=RunConfig(
-                name=self.classifier,
-                local_dir=self._workdir,
-            ),
-            datasets=datasets,
-        )
-
-        training_result = self._trainer.fit()
-        self._model_ckpt = training_result.best_checkpoints[0][0]
-
+    @abstractmethod
     def predict(self, ds):
-        print('predict')
-        # Predict with model
-        predictions = self._make_predictions(ds)
+        """
+        """
 
-        # Convert predictions to labels for cross-validation of classification
-        predictions = self._get_abs_pred(predictions)
+    @abstractmethod
+    def predict_proba(self):
+        """
+        """
 
-        # Return decoded labels
-        return self._label_decode(predictions)
-
-    def predict_proba(self, ds, threshold = 0.8):
-        print('predict_proba')
-        # Predict with model
-        predictions = self._make_predictions(ds)
-
-        # Convert predictions to labels with threshold for top-down classification
-        predictions = self._get_threshold_pred(predictions, threshold)
-
-        # Return decoded labels
-        return self._label_decode(predictions)
-        
-    def _make_predictions(self, ds):
-        if ds.count() > 0:
-            if len(ds.schema().names) > 1:
-                col_2_drop = [col for col in ds.schema().names if col != TENSOR_COLUMN_NAME]
-                ds = ds.drop_columns(col_2_drop)
-
-            # Preprocess
-            if self._scaler is not None:
-                ds = self._scaler.transform(ds)
-            ds = ds.materialize()
-
-            self._predictor = BatchPredictor.from_checkpoint(
-                self._model_ckpt,
-                TensorflowPredictor,
-                model_definition = lambda: build_model(self.classifier, self._nb_classes, self._nb_kmers)
-            )
-            predictions = self._predictor.predict(
-                data = ds,
-                feature_columns = [TENSOR_COLUMN_NAME],
-                batch_size = self.batch_size,
-                num_cpus_per_worker = self._nb_CPU_per_worker,
-                num_gpus_per_worker = self._nb_GPU_per_worker
-            )
-            return predictions
-        else:
-            raise ValueError('No data to predict')
-    
-    def _get_abs_pred(self, predictions):
-        print('_get_abs_pred')
-        def map_predicted_label_binary(ds):
-            ds = np.ravel(ds['predictions'])
-            lower_threshold = 0.5
-            upper_threshold = 0.5
-            predict = pd.DataFrame({
-                'proba': ds,
-                'predicted_label': np.full(len(ds), -1)
-            })
-            predict.loc[predict['proba'] >= upper_threshold, 'predicted_label'] = 1
-            predict.loc[predict['proba'] <= lower_threshold, 'predicted_label'] = 0
-            return {'predictions' : predict['predicted_label'].to_numpy(dtype = np.int32)}
-        
-        def map_predicted_label_multiclass(ds):
-            ds = ds['predictions']
-            pred = pd.DataFrame({
-                'best_proba': [np.max(arr) for arr in ds],
-                'predicted_label' : [np.argmax(arr) for arr in ds]
-            })
-
-            return {'predictions' : pred['predicted_label'].to_numpy(dtype = np.int32)}
-        
-        if self._nb_classes > 2:
-            print('map_predicted_label_multiclass')
-            fn = map_predicted_label_multiclass
-        else:
-            print('map_predicted_label_binary')
-            fn = map_predicted_label_binary
-
-        predict = []
-        predictions = predictions.map_batches(
-            lambda batch : fn(batch),
-            batch_format = 'numpy',
-            batch_size = self.batch_size
-        )
-        for row in predictions.iter_rows():
-            predict.append(row['predictions'])
-
-        return predict
-
-    def _get_threshold_pred(self, predictions, threshold):
-        print('_get_threshold_pred')
-        def map_predicted_label_binary(ds, threshold):
-            ds = np.ravel(ds['predictions'])
-            lower_threshold = 0.5 - (threshold * 0.5)
-            upper_threshold = 0.5 + (threshold * 0.5)
-            predict = pd.DataFrame({
-                'proba': ds,
-                'predicted_label': np.full(len(ds), -1)
-            })
-            predict.loc[predict['proba'] >= upper_threshold, 'predicted_label'] = 1
-            predict.loc[predict['proba'] <= lower_threshold, 'predicted_label'] = 0
-            return {'predictions' : predict['predicted_label'].to_numpy(dtype = np.int32)}
-        
-        def map_predicted_label_multiclass(ds, threshold):
-            ds = ds['predictions']
-            pred = pd.DataFrame({
-                'best_proba': [np.max(arr) for arr in ds],
-                'predicted_label' : [np.argmax(arr) for arr in ds]
-            })
-            pred.loc[pred['best_proba'] < threshold, 'predicted_label'] = -1
-
-            return {'predictions' : pred['predicted_label'].to_numpy(dtype = np.int32)}
-        
-        if self._nb_classes > 2:
-            print('map_predicted_label_multiclass')
-            fn = map_predicted_label_multiclass
-        else:
-            print('map_predicted_label_binary')
-            fn = map_predicted_label_binary
-
-        predict = []
-        predictions = predictions.map_batches(
-            lambda batch : fn(batch, threshold),
-            batch_format = 'numpy',
-            batch_size = self.batch_size
-        )
-        for row in predictions.iter_rows():
-            predict.append(row['predictions'])
-
-        return predict
+    @abstractmethod
+    def _get_threshold_pred(self):
+        """
+        """
 
 
 # Training/building function outside of the class as mentioned on the Ray discussion
