+++ src/models/multiclass_utils.py
@@ -24,43 +24,14 @@ class MulticlassUtils(ModelsUtils, ABC):
     These methods are meant to be used when decomposing data into taxonomic groups before training one model per group
     
     -----------------------
-    Ray data GroupBy
+    Ray data GroupBy + Bagging meta-estimator
     -----------------------
     https://www.anyscale.com/blog/training-one-million-machine-learning-models-in-record-time-with-ray#approach-2:-using-ray-data-(grouping-data-by-key)
+    https://scikit-learn.org/stable/modules/generated/sklearn.ensemble.BaggingClassifier.html#sklearn.ensemble.BaggingClassifier
     1. GroupBy previous taxa
     2. Fx for model training (train_fx)
     3. ds.map_groups(train_fx) to exec the training of models in parallel
     4. Write results to file / save models
-    
-    -----------------------
-    Mixture-of-Experts (MoE)
-    -----------------------
-    1. Train each expert on their task-associated data
-        * Split training data into 80/20% splits
-        * Train/val over multiple epochs
-    2. Train a gating network on the whole task
-        * Perceptron NN for gating
-        * Train on whole training ds
-        * Validation on simulated reads ds
-        * CV on test simulated reads ds 
-    https://medium.com/@bensalemh300/harnessing-the-best-of-both-worlds-how-mixture-of-experts-meets-pyspark-for-mnist-mastery-315f82e65a0e
-    https://machinelearningmastery.com/mixture-of-experts/
-
-    1. Cluster Data Split: Data within each cluster is divided into training and testing sets.
-    2. Decision Tree Classifiers: For clusters where there’s more than one unique class in the training data, we train Decision Tree classifiers. These classifiers can distinguish between different classes within the cluster.
-    3. Storing Expert Models: Trained Decision Tree models are stored in a dictionary, where each expert corresponds to a specific cluster.
-    4. Performance Evaluation: The performance of each expert model is assessed by evaluating its accuracy on the corresponding test data.
-    
-    Sklearn LogisticRegression : https://github.com/zermelozf/esn-lm/blob/master/esnlm/readouts/smoe.py
-    Keras/TF : https://abdulkaderhelwan.medium.com/mixture-of-experts-introduction-39f244a4ff05
-    Keras/TF on article 2018 : https://github.com/drawbridge/keras-mmoe
-    Keras/TF 2018 : https://github.com/eminorhan/mixture-of-experts
-    Detailed example : https://mattgorb.github.io/moe
-    Detailed example : https://towardsdatascience.com/how-to-build-a-wide-and-deep-model-using-keras-in-tensorflow-2-0-2f7a236b5a4b
-    Keras example : https://keras.io/examples/nlp/text_classification_with_switch_transformer/
-    Keras example : https://stackoverflow.com/questions/77551865/how-to-extend-keras-gpt2-model-moe-example
-    FastMoE PyTorch : https://fastmoe.ai/
-    Tutel PyTorch : https://www.microsoft.com/en-us/research/blog/tutel-an-efficient-mixture-of-experts-implementation-for-large-dnn-model-training/
     """
 
     def _get_count_previous_taxa(self, taxa, csv):
