+++ src/models/sklearn/models.py
@@ -93,11 +93,3 @@ class SklearnModels(ModelsUtils, ABC):
     def _get_threshold_pred(self):
         """
         """
-
-    def _label_decode(self, predict):
-        print('_label_decode')
-        decoded = pd.Series(np.empty(len(predict), dtype=object))
-        for label, encoded in self._labels_map.items():
-            decoded[predict == encoded] = label
-
-        return np.array(decoded)
